✨ feat: Add username and password to local credential sharing
- Update CredentialResponse to include username and password fields
- Modify LocalCredentialServer to send user credentials from LoginCredentialManager
- Update LocalCredentialClient response structure to handle new fields
- Enhance CredentialsHandler to use and store credentials from local server
- Add LoginCredentialManager integration for secure credential storage
