+++ KiaExtension/GetCarPowerLevelStatusHandler.swift
@@ -59,6 +59,8 @@ class GetCarPowerLevelStatusHandler: NSObject, INGetCarPowerLevelStatusIntentHan
                         } catch {
                             result = .init(code: .failureRequiringAppLaunch, userActivity: nil)
                         }
+                    case .unexpectedStatusCode(400):
+                        result = .init(code: .success, userActivity: nil)
                     default:
                         result = .init(code: .failure, userActivity: nil)
                     }
