+++ KiaMaps/App/AppConfiguration.swift
@@ -9,10 +9,6 @@
 import Foundation
 
 enum AppConfiguration {
-    /// Your username to Kia/Hyunday/Genesis connect
-    static let username = ""
-    /// Your password to Kia/Hyunday/Genesis connect
-    static let password = ""
     /// Supported brands
     static let apiConfiguration: ApiConfiguration = ApiBrand.kia.configuration(for: .europe)
     /// If nil it will choose first vehicle in list
