+++ KiaMaps/App/LoginView.swift
@@ -287,31 +287,6 @@ struct KiaTextFieldStyle: TextFieldStyle {
     }
 }
 
-// MARK: - Login Credentials Management
-
-struct LoginCredentials: Codable {
-    let username: String
-    let password: String
-}
-
-enum LoginCredentialManager {
-    private enum Key: String {
-        case loginCredentials
-    }
-    
-    static func store(credentials: LoginCredentials) {
-        Keychain<Key>.store(value: credentials, path: .loginCredentials)
-    }
-    
-    static func retrieveCredentials() -> LoginCredentials? {
-        Keychain<Key>.value(for: .loginCredentials)
-    }
-    
-    static func clearCredentials() {
-        Keychain<Key>.removeVakue(at: .loginCredentials)
-    }
-}
-
 // MARK: - Preview
 
 #Preview {
